opportunity to remove some code duplication
