+++ core/src/main/java/com/questdb/std/CharSequenceIntHashMap.java
@@ -23,22 +23,13 @@
 
 package com.questdb.std;
 
-import com.questdb.std.str.NullCharSequence;
-
 import java.util.Arrays;
 
 
-public class CharSequenceIntHashMap implements Mutable {
-    private static final int MIN_INITIAL_CAPACITY = 16;
+public class CharSequenceIntHashMap extends AbstractCharSequenceHashSet {
     private static final int NO_ENTRY_VALUE = -1;
-    private static final CharSequence noEntryKey = NullCharSequence.INSTANCE;
     private final int noEntryValue;
-    private final double loadFactor;
-    private CharSequence[] keys;
     private int[] values;
-    private int free;
-    private int capacity;
-    private int mask;
 
     public CharSequenceIntHashMap() {
         this(8);
@@ -50,25 +41,25 @@ public class CharSequenceIntHashMap implements Mutable {
 
     @SuppressWarnings("unchecked")
     public CharSequenceIntHashMap(int initialCapacity, double loadFactor, int noEntryValue) {
+        super(initialCapacity, loadFactor, 0.3);
         this.noEntryValue = noEntryValue;
-        int capacity = Math.max(initialCapacity, (int) (initialCapacity / loadFactor));
-        capacity = capacity < MIN_INITIAL_CAPACITY ? MIN_INITIAL_CAPACITY : Numbers.ceilPow2(capacity);
-        this.loadFactor = loadFactor;
-        keys = new CharSequence[capacity];
         values = new int[capacity];
-        free = this.capacity = initialCapacity;
-        mask = capacity - 1;
         clear();
     }
 
     public final void clear() {
-        Arrays.fill(keys, noEntryKey);
+        super.clear();
         Arrays.fill(values, noEntryValue);
-        free = this.capacity;
     }
 
-    public boolean contains(CharSequence key) {
-        return keyIndex(key) < 0;
+    public boolean removeAt(int index) {
+        if (index < 0) {
+            Unsafe.arrayPut(keys, -index - 1, noEntryKey);
+            Unsafe.arrayPut(values, -index - 1, noEntryValue);
+            free++;
+            return true;
+        }
+        return false;
     }
 
     public int get(CharSequence key) {
@@ -82,38 +73,6 @@ public class CharSequenceIntHashMap implements Mutable {
         } else {
             putAt0(index, key, 0);
         }
-    }
-
-    public CharSequence keyAt(int index) {
-        return index < 0 ? Unsafe.arrayGet(keys, -index - 1) : null;
-    }
-
-    public int keyIndex(CharSequence key) {
-        int index = Chars.hashCode(key) & mask;
-
-        if (Unsafe.arrayGet(keys, index) == noEntryKey) {
-            return index;
-        }
-
-        if (Chars.equals(key, Unsafe.arrayGet(keys, index))) {
-            return -index - 1;
-        }
-
-        return probe(key, index);
-    }
-
-    public int keyIndex(CharSequence key, int lo, int hi) {
-        int index = Chars.hashCode(key, lo, hi) & mask;
-
-        if (Unsafe.arrayGet(keys, index) == noEntryKey) {
-            return index;
-        }
-
-        CharSequence cs = Unsafe.arrayGet(keys, index);
-        if (Chars.equals(key, lo, hi, cs, 0, cs.length())) {
-            return -index - 1;
-        }
-        return probe(key, lo, hi, index);
     }
 
     public boolean put(CharSequence key, int value) {
@@ -146,50 +105,8 @@ public class CharSequenceIntHashMap implements Mutable {
         }
     }
 
-    public boolean remove(CharSequence key) {
-        return removeAt(keyIndex(key));
-    }
-
-    public boolean removeAt(int index) {
-        if (index < 0) {
-            Unsafe.arrayPut(values, -index - 1, noEntryValue);
-            free++;
-            return true;
-        }
-        return false;
-    }
-
-    public int size() {
-        return capacity - free;
-    }
-
     public int valueAt(int index) {
         return index < 0 ? Unsafe.arrayGet(values, -index - 1) : noEntryValue;
-    }
-
-    private int probe(CharSequence key, int index) {
-        do {
-            index = (index + 1) & mask;
-            if (Unsafe.arrayGet(keys, index) == noEntryKey) {
-                return index;
-            }
-            if (Chars.equals(key, Unsafe.arrayGet(keys, index))) {
-                return -index - 1;
-            }
-        } while (true);
-    }
-
-    private int probe(CharSequence key, int lo, int hi, int index) {
-        do {
-            index = (index + 1) & mask;
-            if (Unsafe.arrayGet(keys, index) == noEntryKey) {
-                return index;
-            }
-            CharSequence cs = Unsafe.arrayGet(keys, index);
-            if (Chars.equals(key, lo, hi, cs, 0, cs.length())) {
-                return -index - 1;
-            }
-        } while (true);
     }
 
     private void putAt0(int index, CharSequence key, int value) {
